+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Implement @EnableFuzzyRepositories
 - Support mapping java.util.Date into a fuzzydb record
+- Support mapping org.bson.types.ObjectId into a fuzzydb record if on classpath
 - Add fuzzydb.tld function taglib to ease use of displaying results